// frontend/src/app/core/interfaces/phones.interfaces.ts
// Generated by https://quicktype.io

export interface PhonesResponse {
  data: Data;
  meta: Meta;
}

export interface Data {
  id: number;
  attributes: Attributes;
}

export interface Attributes {
  name: string;
  area_code: string;
  number_phone: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  primary: boolean;
}

export interface Meta {}

// Generated by https://quicktype.io

export interface PhoneBody {
  id: number;
  area_code: string;
  number_phone: string;
  name: string;
  primary: boolean;
}